Use svg impages instead of plantuml sphinx plugin
PlantUML is good tool for developers to ease create diagrams
but it uses Java to work.
Moving development docs to readthedocs is easy but it doesn't
allow us to use Java plugins for Sphinx.
We'll continue to use PlantUML for diagrams generation but
we need to store generated images for readthedocs hosting.

Change-Id: I9cc44d4fa8a5055768938011e5f599740a8b989f