Remove Network and Vlan models
These models are removed and NetworkGroup model have all relationships that Network model had.
Vlan model was utilized to check that newly created networks
have the same VLAN IDs being assigned as networks of other clusters.
Now this can be checked using NetworkGroup model but not checked
because the result of this checking was used for logging only.

Implements: blueprint nailgun-remove-network-model

Change-Id: Iaec7d5acdb0e47d9e9cf4c224e1be768df83107f