Flexible networks-to-nics assignment
Networks-to-nics assignment is now sensitive to current networking configuration of target cluster.
Default assignment in case of Nova-Network is changed because the current algorithm tries to map
as more as possible networks to each NIC
instead of fixed mapping to non-admin NIC of all remaining networks as it was done before.

Closes-Bug: #1262149

Change-Id: Idd62f124b0f12ac10f3b542cdb05a476ec36cfec