Allow multiple networks to be created per cluster
Added a NodeGroup model. NodeGroups can be created and modified
via the Fuel API. Clusters are created with a default node group
and the cluster networks are tied to that default node group.

Upon creation of a node group a set of networks will be generated.
When a node group is deleted those networks will be deleted as well.

See previous change: Ic30a1b46112599022645b06f7bef2223ff4c6475

Change-Id: I8bbe6b8c6338d6586c7e9a61542939ae2a19828b
Implements: blueprint multiple-cluster-networks