Don't update NIC info, if metadata is incorrect
Sometimes our discovery agent tries to post a new metadata with missed NICs.
In that case, nailgun used to remove NICs from the database, even those with
assigned network roles. This leads to bad consequences, such as an exceptions
in serialization phase and so on.

Since now, if a NIC has assigned network roles, it won't be removed from
the database, even if a new metadata has been received without that NIC.
So we just log a warning message and print a traceback to logs.

Closes-Bug: #1282568
Change-Id: I5658379587d2adec4b5baa0419aca4659bdb7883