Set minimal memory limit for CirrOs image to 64MB
TestVM (that is cirros) is installed by fuel
in each new openstack cloud. This image doesn't
have any limits set in openstack. If user tries
to launch this image on flavor with 16Mb of
memory it will go into infinite boot. It would be
good to prevent such accidental hard-to-find
issues by setting limits for image on required
resources.

Change-Id: Ifced26e747e1a688d1f0e8026c447da425cdf86e
Related-Bug: #1304664