Tries to get NIC by name if mac was changed
Fuel loses nodes when some node was added to cluster and then mac-addresses
of NICs have been changed. Since now the issue is fixed.

The fix is based on the assumption that a NIC name is unique within a
node, and if we can't get an interface by MAC, we can try to get it by
name.

Closes-Bug: #1282568
Change-Id: I612a183d51ebb00afbf753ae3ee8dbccac54a4c4