[UI changes] Added ability to create raid of UI
- Added a new tab Storage, which shows the status of all
controllers on all nodes of our environment. For each
controller, we can view all created virtual disks (raid),
all created caches, global hot spare and unused drives.
As well as state of all members of these disk groups.

- On the disk configuration added sub tabs. When
selecting a node that has a raid controller, on this
page we can go to the tab controller. Where we can see
the virtual drives, if they are created, and unused
drives. For unused disks available opportunity to select
a number of drives and convert them to a new virtual
drive. In this type and their amount depends on which of
them we can create.

- Added new dialogs: create virtual drives and warning
about possible data loss when you press reset, revert
changes and load defaults on the configuration page raid
controllers.

Change-Id: I6901f14f67c51d29e69adac03ed9d929fe0ca828