Mark nodes to error if update was failed
We have to mark all nodes to error explicitly in case of failed cluster
updating. Such behaviour was implemented for regular deploy process
but was mistakenly omitted for the update process.

Please keep in mind, that all nodes are marked as error only in case
Astute didn't sent per-node status information.

Change-Id: I945c96d8fc7a27cb7779894a1f69baaafd096689
Closes-Bug: #1332494