Acceptance tests for cluster and node operations
Tests with linear growing of number of nodes in the environment.
Each operation is measured with growing nodes number.
In tests checks execution time and operation speed degradation.
Tests are not passed, due to low speed of some operations.
This set of tests should be used as part of acceptance tests for fuel-1k blueprint.

Change-Id: Ic747e8e9781393a0cc076e1e3f6f3d35db1c4f65
Blueprint: fuel-1k