Enable signal trap for fuelmenu to save and quit
If a user sends kill -SIGUSR1 to fuelmenu, it
will attempt to save all settings (with a 60 second
timeout) and quit normally. If the timeout expires,
changes are not saved and the program exits.

This operation mode is ideal for system tests where
applying custom Puppet manifests affects Fuel Master
deployment.

blueprint fuel-master-ci-tests

Change-Id: Iae40dc73f4133266f084c2a625e54220600125d7