Use masks in release orchestrator_data
Since Fuel Upgrade feature, it's a common practice to add new
openstack release, so it'd be very useful to add new release
with a kind of macro replacement in paths to repos and puppets.

Since this commit, the macro replacement for MASTER_IP was
introduced.

(cherry picked from commit 2488bffd55baee84b33366ac9e4eef62dbd70607)

Closes-Bug: #1378901
Change-Id: Ie3a79ab105461f10cbabcc5353ca952946200f32