Removed sleep in execute util in fuel-agent
Sometimes when parted performs some action
it fails to re-read updated partition table
on a disk. It occurs mostly when you run parted
command several times without delay. We had
sleep(1) in execute utility as a fast workaround.

Looks like the issue can be addressed with just
additional call or partprobe right after failed
parted command. This patch impelemnts this
assumption.

But it is still not a real fix, it is again just
a workaround. Real fix is to launch parted one time
for every disk passing to it a whole table
(all partition at once). We are going to implement
this as soon as possible, right after addressing
most prioritized issues.

Change-Id: I539800271260c3d3f1b3560f8111ba8b6cacc9c4
Implements: blueprint image-based-provisioning