Basic profiling facilities for CLI
This patch implements a simple wrapper for
profiling different CLI commands and saving
the results.

cProfile is used as a profiling tool; results
are saved as a call-tree, textual statistics and
a DOT-graph.

In order to enable profiling NAILGUN_PERFORMANCE_PROFILING
environment variable should be set to one of 1, True or Yes.
The value of this variable is case-insensitive.

DocImpact
Blueprint: 100-nodes-support
Change-Id: Ia9f6d19960fbe35e034fa85bf6303f3a419a7a75