Basic profiling facilities for CLI
This patch implements a simple wrapper for
profiling different CLI commands and saving
the results.

cProfile is used as a profiling tool; results
are saved as a call-tree, textual statistics and
a DOT-graph.

Performance profiling settings are stored in settings.yaml
which is introduced by this patch and is distributed with
Fuel Client. Custom user settings can be supplied using the
FUELCLIENT_CUSTOM_SETTINGS environment variable overriding
the default ones.

DocImpact
Blueprint: 100-nodes-support
Change-Id: Ia9f6d19960fbe35e034fa85bf6303f3a419a7a75