Fix fuel_6_0.py migration script
We have wrong approach of writing alembic migration scripts -
we have all-in-one migration file, which is wrong, because Alembic
itself promotes to use a separate file for separate migration task.

The issue shows us that we can't and shouldn't separate schema and data
migration, since they are a whole thing and rely on each other.

Change-Id: Ic1c5c69243211ae9b48088b6a4787a47ce6e5515
Closes-Bug: #1391525