In case of several plugins, nailgun should use the newest
For example Fuel has 3 plugins:
- name: plugin_name, version: 1.0.0
- name: plugin_name, version: 2.0.0
- name: plugin_name, version: 0.1.0

The problem was, when user enables a plugin on UI,
Nailgun enables all of the installed plugins.

With this patch, after user creates environment he can
enable or disable only 2.0.0 version of plugin.

If after environment creation, new version of the plugin
is installed, user cannot use this new version, he
should create new environment in order to enable it.

Change-Id: I51ba03166cd2a7379c6f0912d668265bf88a316d
Closes-bug: #1395342