Rename Cinder role
The name and the description of the storage role in openstack.yaml
was hardcoded to mention that this role provides block storage over
iSCSI and uses LVM. That is not true for cases when VMware vCenter
is used as a back-end for storage.

This patch changes those name and description to more common ones.

DocImpact
Closes-bug: 1383224
Change-Id: I4aa013611242ff6b7e6e0889b38bfbf9e9a700ec