Replace volume containers with direct mountings
Fuel storage containers are a bit useless in our architecture and lead
us to the problems when after upgrade we have storage containers with
not expected names. In order to be consistent with dockerctl and prevent
desync with dockerctl it was decided to remove them.

(cherry picked from commit b5171a5797b46303b3916abe946766a35d0a3f16)

Change-Id: Id3b614ce5f8cd5284bc387d1384d0c9928f84981
Related-Bug: #1382527