urllib3 exceptions are logged without trace
urllib3.exceptions.ProxyError and urllib3.exceptions.DecodeError are
the urllib3 exceptions which are not handled by requests (old version)
so it is required to catch urllib3 exceptions separately.
This can be removed when requests==2.4.3 will be included to Fuel
(see https://bugs.launchpad.net/fuel/+bug/1394182).

This patch requires changes to fuel-main:
https://review.openstack.org/137343

Closes-Bug: #1390457

Change-Id: I30891796ea85289f81231c4cad3c802e33d21b03