Return error HTTP code for incorrect neutron conf
Recently, we used to return error message for incorrect neutron
configuration without error HTTP status code. In other words, the
"202 Accepted" was always returned regardless it was successed or not.

Since now, tbe "400 Bad Request" will be returned in case of error and
"200 OK" in case of success (there's no reason to return "202 Accepted"
for synchronous task).

Change-Id: I9ad16fceb930332b427773dec32bd84d30fb1fd3
Closes-Bug: #1404821