Move graph creation outside of serializaer creation
DeploymentSerializers should be treated only as config serializers,
and should not contain logic that is related to tasks and order of
execution.

Right now we are bound to use graph serializer inside deployment serializer
because of current reference architecture.

Also graph will be used as a base for pre/post tasks serialization

implements blueprint granular-deployment-based-on-tasks

Change-Id: Ice986f778a230ebf69c16eb074543a5d8afda672