fuelclient: use keystone v2.0 client explicitly
Currently, we're using a Keystone client which actually performs
auto discovering and returns one of available client. In theory
all works fine, but unfortunately the "auth_url" will be replaced
by Keystone with a keystone endpoint. The issue is that this
endpoint may be unavailable if we're working through some
proxy/tunnel and don't have direct access to the fuel master node.

As workaround, we can explicitly create a keystone client for v2.0,
and none of discovering will be performed and the endpoint will not
be replaced implicitly by keystone client.

Change-Id: Ieada040739116227f53f70558cead5bcb4d5bb48
Closes-Bug: #1405190