Move fuel-web docs out, keeping only API docs
This commit removes the documentation contents
from this repository, as they are being moved
into fuel-docs repository. (link to CR will be
added here when available)

The autogenerated API docs will still be built
in this repo, and they will be referenced from
the main Fuel docs.

Change-Id: I5a06bdaca39127e23c89ca9ff45e50f8e16a19fe
Implements: blueprint fuel-dev-docs-merge-fuel-docs