Introduce RootComponent
This component will be a root component, so it will remain unchanged
even if the page (which was previous top-level component/view)
is changed. This change is needed to complete rewriting ClusterPage
to React as it has subcomponents with transition which cannot survive
remounting of the top-level component.

Related to blueprint react-router
Related to blueprint backbone-to-react

Change-Id: I4b44ad46e1cd491d751958a082c732ffaed04563