Reform releases '.state' attribute to fit our goals
Current release's '.state' attribute is obsolete and do not used. It was
introduced a time ago when we had an RHEL support. So it could be
refactored to fit out current goals.

Since now there will be only two states:

* available
* unavailable

There is no way to start provisioning task for unavailable releases.

Change-Id: I8827c710da6e4f77a6397381d51fe25b87149462
Implements: blueprint downloadable-ubuntu-release