IBP: ubuntu: moved resolv.conf stuff to boothook
The issue is that cloudinit/cc_resolv_conf.py is only able
to deal with plain /etc/resolv.conf file and does not know about
resolvconf and sysconfig mechanisms which allow user to
generate this file automatically. The problem is currently
actual for ubuntu only, so, I moved this resolv.conf stuff
for ubuntu on a boothook stage. For Centos we just edit
/etc/resolv.conf file.

Closes-Bug: 1416399
Change-Id: I3d6fdd15ea279928d39f5e6ce4136372db3172c5