[IBP] Fix udev 70-net-persistent.rules filling
It's related to deployment snippents backporting to cloud-init.
Snippet works with kernel parameter 'udevrules' passed during booting to
/proc/cmdline.

Image based installation doesn't use /proc/cmdline for getting this data.
And it was containing redundant for that case 'grep udevrules' in the scripted
line. It was creating empty file for every time. After removing it, the issue
becomes completely resolved.

Change-Id: I0fb3c5bb99960c71bb4e9299da144ea9ca7479fc
Closes-Bug: #1419769