Add mode for fake_tasks, fixes hanging task
demigodmode is added, it makes all tasks created with fake_tasks using
this option fix themselves on 50 progress and 'running' status.

test_rerun_stopped_deploy uses above option for first deploy, which is
meant to be stopped, but couldn't be run in a manner that guarantees
that stop task will be fast enough to stop the deploy.

Change-Id: Ie00e3845b8bb2273ce9857b789071705c81d2b34
Closes-Bug: #1419046