Data format of workload resource_data changed
In case of cluster reset we can lose information about adding and
removing workloads. For prevention format of resource_data is
changed. Now we have lists of dicts in 'added', 'remove', 'modified'
and 'current'. On cluster reset we can have several records in the
'added' with different time of adding. For 'removed' we add full
data only if it doesn't already added. In other case we add only
id with removing time.

Change-Id: I763705b196ec78aa9ed2df1a3fc8f98325377ae0
Partial-Bug: #1427757
Implements: blueprint openstack-workload-statistics