ClusterChangesHandler task called asynchronously
We use uWSGI mules for this task. This is because on scale lab we have timeout
errors which are reported like this:

DEBUG - Uuuups https://bugs.launchpad.net/fuel/+bug/1384623 again

Thus the idea to offload most of the work done in this handler to an
asynchronous worker (uWSGI mule in this case). Note that increasing uWSGI
workers count doesn't help -- timeouts for this specific request will still
occur, we would only have the advantage of not blocking the API so much.

Other suggestions were made:
- use receiverd -- this has the disadvantage of breaking current flow of
  messages which is something like Nailgun -> Astute -> receiver
  (we would have in this case: Nailgun -> receiver -> Astute -> receiver
  which is not intuitive)
- use Celery -- this is a big change and not a bad idea for the future
  but we would like to have this merged for the 6.1 release, mules are at
  hand already
- performance improvements -- this is not that easy too (as noted in the
  blueprint 200-nodes-support)

blueprint: 200-nodes-support
Related-Bug: #1428957

Change-Id: Ib456efe434e956adfca7d20eb4a831e1c6102342