Added download-debian-installer script
We are not allowed to re-distribute ubuntu debian-installer
kernel and initrd. So, we are going to download these files
from one of user defined ubuntu repos. We implemented this
as a pre-provision task which is supposed to run a shell
script.

This patch introduces this script for downloading
debian-installer initrd image.

Change-Id: I9e7ac50d19ea8b984e1ad2831c73f2cb88ddbd54
Implements: blueprint consume-external-ubuntu