Updating functionality for action logs changed
The main reason this patch is introduced - issue with two Nailgun tasks:
check_networks and verify_networks which are constructed from the same
instance of task object. This in turn leads to dublication of particular
attributes in action_logs entries and so affects update routines for
them. Also check_networks is not updated to status "ready" in case of
success so action logs processing methods are changed to deal with this
issue.

Change-Id: I949eec203b629ef6d1d0d6664540c404dc12c36c
Closes-Bug: #1430812