Add support for 'gateway' option provided in settings
The 'gateway' field is used by NetworkManager.get_default_gateway()
in the Nailgun. For 'fuelweb_admin' network it was empty, so
settings.MASTER_IP was used as default gateway.

Now, default gateway for 'fuelweb_admin' network can be configured
in the /etc/fuel/astute.yaml, option dhcp_gateway; so this setting
should be used as default gateway for provisioning.

Related to blueprint consume-external-ubuntu

Change-Id: I36d718c62f40d36dd3260038d087d630d45e5f34