Fix error with no updating of action log after empty cluster deletion
In case user deletes empty cluster through api, action log is not updated
for cluster_deletion task as orm related error occurs in updating routine.
The reason for it is that we try to operate on task instance which
corresponding row was removed from db and hence attempt to access
attributes, marked on model as deffered, produces mentioned error.

This patch wraps getting of task.cache attribute into separate
try..except block in which case we only lose update for "output" section
of additional_info for action_log entry.

Unit test for implemented task helper routine, modeling situation with
referring to "deffered" attribute of instance, which corresponding db row
was deleted, added.

Integration test that emulates user's action leading to the bug added.

Change-Id: Id296c403423e314dd88063883141aff661c7aa9e
Closes-Bug: #1407716