Use `archive` in apt pinning instead of `codename`
We have to use `archive` name in APT pinning because `suite` is actually
an archive name. For instance, the following upstream Ubuntu repo

    deb http://archive.ubuntu.com/ubuntu/ trusty-updates main

has the following attributes in APT pinning

    a=trusty-updates, n=trusty, c=main

Obviously, it's wrong to use `n=` (codename) for matching suite.

Besides this, since now we're generating a few pinnings for the source
like

    deb http://archive.ubuntu.com/ubuntu/ trusty main restricted

because the following line

    a=trusty, c=main, c=restricted

will apply pinning only for the `restricted` section.

Implements: blueprint consume-external-ubuntu
Change-Id: Ic9551ba2da6cacee1ee2f623606eae05342a2ac0