Add parser for roles in dump config.
Add ability to specify all roles in the DUMP variable of nailgun settings.
That makes it possible to specify commands to execute only on the nodes
with selected roles.
Before only controllers and slaves were distinguished.
Also, role name can be specified with filters:
".random" to filter one random node withing the role
".first" to filter first node in sorted by fqdn list of node with the role
".last"  to filter last
E.g. "controller.random", or "ceph-osd.first"
Also, "master" and "local" sections of the tmeplate can be not present in the template.
Before they were mandatory.

Change-Id: I4e23b21cf276c476cf13a93a8a6c52230d4faa1b
Related-Bug: 1420054