Fix url joining in make_download_debian_installer
Unfortunately, we can't use 'urljoin' in make_download_debian_installer
task, becayse it behaves bad in case when input 'uri' doesn't have a
trailing slash. For example, the following line

    >>> urljoin('http://myrepo.com/debian', 'dist/jessie')

will produce the following output

    'http://myrepo.com/dist/jessie'

which is obviously wrong, since we have missed the 'debian' part of the
uri. As alternative, we can use 'os.path.join' here, since we're
intended to be used on POSIX platforms.

Closes-Bug: #1436962
Change-Id: I1d326f12e8082d904541102489b0e31e1e665f2c