Shotgun will always try to save stdout and stderr
Added CCStringIO - "carbon copy" StringIO which in addition
to normal behavior can pass any write to different file-like
object.
It is used as a stdout for executing commands via fabric
and when errors occur during command execution it is used as a source
of stdout. In addition it also proxies everything to
regular sys.stdout.
Removed capturing of stderr for fabric's "run" since
it is always in stdout in our case

Change-Id: I022128141ed142561213a9633bc0fe0f8ccc2685
Closes-Bug: #1434574