Shotgun will always try to save stdout and stderr
Added CCStringIO - "carbon copy" StringIO which in addition
to normal behavior can pass any write to different file-like
object.
It is used as a stdout and stderr for executing commands via fabric
and when errors occur durring command execution it is used as a source
of stdout and stderr. In addition it also proxies everything to
regular sys.stdout and sys.stderr.

Change-Id: I022128141ed142561213a9633bc0fe0f8ccc2685
Closes-Bug: #1434574