Fix issue with non-determenistic ordering in networkx sorting
In current master of networkx it is possible to specify fabric
that will be used for internal graph data structures.

To avoid non determinitic issues with ordering we need to use
OrderedDict that will return same order of keys even with different
hash seeds. OrderedDict should be used for storing order of
nodes inserted in graph, and for dependencies between nodes
in graph.

Changed tests to verify that order will be same with different
hash seeds.

Networkx 1.9.1 (one which we are using)

  http://bit.ly/1CcgdBs

Networkx change with fixed code

  http://bit.ly/1al6Tpq

Change-Id: I16a880fe996a03645c0f79e3156ae403b9f006f5
Closes-Bug: 1438203