Add possibility to change repos on deployed envs
Despite the fact that repos are part of cluster attributes and we can't
change cluster attributes after deployment, we have to be able to change
repos list. It makes sense, because for instance, user wants to add new
nodes but his mirrors were moved to another server, so current repos
are unavailable.

Implements: blueprint consume-external-ubuntu
Change-Id: Ic1a3b721572397f2ee3b030934018fe940b91370