Add 'always_editable' attribute to repo_setup
Despite the fact that repos are part of cluster attributes and we can't
change cluster attributes after deployment, we have to be able to change
repos list. It makes sense, because for instance, user wants to add new
nodes but his mirrors were moved to another server, so current repos
are unavailable.

This patch adds support of the 'always_editable' flag in openstack.yaml.
With that flag we can mark attribute groups to be available for editing
after deployment.

Related-To: blueprint consume-external-ubuntu
Change-Id: I0c58a03b2cd83e1fb196f5a55002d5a869277c6f