IBP Got rid of md for /boot
Centos does use legacy grub, and it is
only able to boot from md with metadata 0.9.
So, we are limited to have not more than 28 disks
on a node because current version of volume
manager assumes /boot is spread over all disks and
0.9 metadata does not support more than 28.
To avoid this limitation we've got rid of using
md for /boot completely.

Change-Id: I08398453625a4e9136d67989c7ebea41cb9cb766
Closes-Bug: #1340414