Implement advanced apt pinning
Current approach is an error prone, because Debian repos may contain
metadata that doesn't fit our pinning template. Yes, in most cases
our template will work, but we want to cover all possible cases.

In order to set corrent pinning, we have to download Debian's Release
file, parse it and create pinning based on it. In that case we'll use
all available pinning rules, so all possible issues will be up to repo
maintainer and not because we're using error prone approach.

Related-Bug: #1435892
Related-Bug: #1438783

Change-Id: I5279b7f414f831503fd993109dd5c91bef51a9ef
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>