Fix for fetching actor_id for action log
Actor_id is required for proper constructing of the action log.
Actor_id value is added into web context environment in the connection
monitoring middleware. When we perform async task processing in the mule,
web request context and environmet is lost. So action_log couldn't be
constructed properly.
Extraction of actor_id from action_log of parent task is added as fix.

Change-Id: I5f0f78073f514d9600cd53d5bf757f27b8bfabe6
Closes-Bug: #1442990