Check bonds instead of their slave interfaces
This allows to verify bonds for deployed nodes and changes verification
algorithm for non-deployed ones.
Now bonds are passed to net_checker for deployed nodes, so it is able
to check the connectivity on bonds when they are configured
(traffic cannot be seen on slave NICs in this case, at least for linux bonds).
For non-deployed nodes slave NICs of bonds are passed to net_checker.
It is checked that at least one of slave NICs of every bond has received
test data.

Additional info (node status, node name, node bonds names and slave NICs) is
also saved into task's message to be able to process net_checker results even
when some nodes were deleted from DB during the verification. It was made to keep
compatibility with previous approach.
We could lock nodes for deletion while they are being checked but it is subject
for another discussion.

DocImpact
Closes-Bug: #1321765
Closes-Bug: #1438945
Change-Id: I94a1beacd4c55867a5a1bfb03ac29e21eded43dc