Exploit default OS credentials for clusters w/o workloads_collector user
Due to service user "workloads_collector" is not created for clusters
with older version than 6.1 after upgrade of master node, oswl
collecting procedure fails for those clusters. This fix adds workaround
to the problem by falling back to default credentials for the cluster to
be used for OS clients initialization. Please, be advised that such
approach will not work in case the user changed the credentials for the
cluster.

Change-Id: I65e9029fec3055a1e1f1179d77696b0338743663
Closes-Bug: #1444410
Closes-Bug: #1444362