Correct APT pinning
In order to set corrent pinning, we have to use repo's metadata and
that means download repo's Release file, parse it and create pinning
based on it. In that case we'll use all available pinning rules, so
all possible issues will be up to repo maintainer and not because
we're using error prone approach.

Port-Of: https://review.openstack.org/#/c/175358/

Co-Authored-By: Igor Kalnitsky <ikalnitsky@mirantis.com>
Related-Bug: #1438783

Change-Id: I71f2b72d914a39bbdf3f094e6b8fdd407fdc725b