Fedora kernel install in ibp mode
Nailgun puts 'kernel_lt' flag in provisioning data and
if this flag is set, it is supposed that we need to install
kernel-lt package which is fedora kernel 3.10.
We assume both kernels are available in centos OS image:
2.6 and 3.10. We can just configure grub to boot with
one of those kernels. If flag 'kernel_lt' is NOT set, we
use 'vmlinuz-2.6' regexp for looking up for kernel file.
Otherwise we use default regexp 'vmlinuz' and sort available
kernels in backward direction, i.e. we use the newest available
kernel.

Co-Authored-By: Vladimir Kozhukalov <vkozhukalov@mirantis.com>
Change-Id: I1f7eee934440ce32d6e733c417e82f578b6d0c18
Partially-closes-bug: #1398643