Add re-tries to Nailgun client
Sometimes, uWSGI may be shutdown for a while because of various reasons
and we should retry our requests to Nailgun and do not fall immediately.

For example, our service checker may report success to early after
Docker upgrades, because Nailgun's puppet run supervisor for a short
time (checkers get it and reports success), then start.sh stops it
(http requests to Nailgun will fail) and runs supervisor as daemon.

Closes-Bug: #1451888

Change-Id: I0e65c346181e6d5d9b17471b02561afa5abbb871
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>