Check master node connection to repositories
Added CheckRepositoryConnectionTask which tries to retrieve files from
repositories and fails if it cannot do it. This check is performed
before provisioning takes place, so it covers the IBP need
for master node to have access to repositories for building an image.

Above task is mocked for integration tests.

Change-Id: I1af89833e7de8447e73b9d9f00dcc2d9170ea35f
Partial-Bug: #1439686