Creating keystone user on customer stats opt-in
For creating keystone user on customer stats opt-in we are using
astute 'execute_task' with type 'puppet'. Puppet granule
'workloads_collector.pp' used for it. Puppet manifest will be applied
only on primary-controllers of operational clusters.

If customer is opted-in to stats before cluster deployment - keystone
user will be created during deployment process without additional tasks
in the Nailgun.

On the opt-out action nothing is to be done by design.

Items of the change:

- Creating tasks on user creation added to the master node settings handler
- Task name 'create_stats_user' added
- Method for getting primary node added to the objects.Cluster
- Task CreateStatsUserTask for executing puppet manifest for creating stats user added
- Fake thread for processing astute 'execute_task' added
- Logging of the task CreateStatsUserTask into action_logs handled
- Extracted base test class for master node settings tests

Change-Id: I7186bb485969c57789cd1e1cc588a8f7b8ba9917
Closes-Bug: #1452821