Update ceph wizard page to controll all four options
Previously we only configured ceph for glance, or cinder, but this
excludes ephemeral or objects. Frequently deployers find this out
the hard way.

This changes the storage wizard page so that we either turn on all
ceph options or leave them all off (the default) and gives some improved
details about the impact of turning it on.

Change-Id: I73aeef0e6c7bb7d123d2086c90ba3d7289c0dc89
Closes-bug: #1454398