Send node_deletion task within other subtasks
We have to execute node deletion task only when provision, deployment
and other tasks are in the database. Otherwise, it may be executed
too quickly (e.g. our tests) and this will affect parent task calculation:
it will be marked as 'ready' because by that time it have only two
subtasks - network_check and node_deletion - and they're ready.

As a fix, we could send task to execution when all others tasks are
in the database.

Closes-Bug: #1455064

Change-Id: I78c45959f9253ee73460259458c46c71f78434dd
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>