Re-raise exception in fake threads
Sometimes our tests fail due to exceptions in fake threads. In this case
we handle exceptions, do rollback and raise it again. The behavious is
wrong, because when we raise this exception again the original traceback
is lost. In order to avoid this we have to re-raise exception. In could
be done by replacing "raise e" with just "raise".

Closes-Bug: #1456161

Change-Id: I960fa1b90ddefe1c9f2936ce90758405eb871ef9
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>