Fix test test_cluster_locking_after_deployment
This fix does not solve the root cause of the problem,
deadlock with fake thread is still there, but proper
solution with correct lock ordering is huge, so we
won't be able to land it in 6.1, this patch fixes
the problem and lives logic of the tests as is.
As result it will be possible to make backports to
this branch without random test fails.

Change-Id: I7cf4c4ce0507afeda760fc1fbdeabac43b0e2909
Closes-bug: #1455467