Add controller to deployment critical roles list
Currently if a failure happens on a controller as part of an HA
deployment, the failure is ignored and the deployment continues on the
rest of the nodes. This can lead to errors in the post deployment
phase. One particular instance is when Ceph is being utilized and
we attempt to restart the rados processes on all of the controllers.

In order to prevent errors that may occur in the post deployment
phase, we need to ensure that all actions were successful on all of
the regular controller nodes as well as the primary-controller role.

Change-Id: I466c9f13f373d9a7a5e05e365c173895c87a56a2
Closes-Bug: 1456318
Related-Bug: 1455570