Restart mcollective at the end of cloud-init magic
We use mcollective cloud-init module to configure
mcollective service. This module runs the command
'service mcollective start' which does nothing
if the service is already running (which is the actual case).
So, we can end up with mcollective unable to see
config file changed made by cloud-init and hence
prevent mcollective from being able to connect to
rabbit.

This patch restarts mcollective at the end of cloud-init run.

Change-Id: I9ea122acc168120596bf48f378e92874936bcbe1