Use netboot installer from trusty-updates
It makes sense to use a netboot installer from trusty-updates repo,
not from trusty, because it provides more fixes and updates. Moreover,
the `fuel-createmirror` downloads exactly netboot installer from
trusty-updates and ignores trusty's one. That basically means that
currently classic provisioning doesn't work with partial mirrors at all
an this commit is going to fix that.

Closes-Bug: #1461927

Change-Id: I265b212bcd7aae9b8f6764806da65b8cb98312b9
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>