Set upper bounds of packages using old pbr
Because some of the packages (python-cinderclient, keystonemiddleware)
do not release new versions with pbr requirements aligned with
OpenStack global requirements, we have a conflict in requirements
since new packages install pbr 1.0.1 (newest one) and "old" packages
want pbr<=1.0.
Because of that we limit higher version of explicitly installed packages
so we avoid this issue.

Change-Id: If90e0795cf8cebdeb381b17ad0a694b7b35f3d88
Closes-Bug: #1461980